Make Anchor a newtype and move it to Types
